added (and commented out) tscondig for build and dev env
