+++ lib/components/Keyboard/KeyboardInput/utils/EventEmitterManager.ts
@@ -1,6 +1,7 @@
 import _ from 'lodash';
 
 export default class EventEmitterManager {
+  handlerCallbacks = {};
   constructor() {
     this.handlerCallbacks = {};
   }
