+++ lib/components/Keyboard/KeyboardTracking/KeyboardTrackingView/KeyboardTrackingView.ios.tsx
@@ -3,34 +3,11 @@
  */
 
 import React, {PureComponent} from 'react';
-import ReactNative, {requireNativeComponent, NativeModules, ViewStyle} from 'react-native';
+import ReactNative, {requireNativeComponent, NativeModules} from 'react-native';
+import {KeyboardTrackingViewProps} from './index';
 
 const NativeKeyboardTrackingView = requireNativeComponent('KeyboardTrackingViewTemp');
 const KeyboardTrackingViewTempManager = NativeModules.KeyboardTrackingViewTempManager;
-
-export type KeyboardTrackingViewProps = {
-  /**
-     * Enables tracking of the keyboard when it's dismissed interactively (false by default).
-     * Why? When using an external keyboard (BT),
-     * you still get the keyboard events and the view just hovers when you focus the input.
-     * Also, if you're not using interactive style of dismissing the keyboard
-     * (or if you don't have an input inside this view) it doesn't make sense to track it anyway.
-     * (This is caused because of the usage of inputAccessory to be able to track the
-     * keyboard interactive change and it introduces this bug)
-     */
-   trackInteractive?: boolean;
-   /**
-    * Allow control safe area
-    */
-   useSafeArea?: boolean;
-
-   // Can't figure out what it's supposed to be
-   ref?: any;
-
-   children?: React.ReactChild | React.ReactChild[];
-
-   style?: ViewStyle;
-}
 
 /**
  * @description: A UI component that enables “keyboard tracking" for this view and it's sub-views.
