+++ lib/components/Keyboard/index.ts
@@ -1,7 +1,15 @@
-import KeyboardTrackingView from './KeyboardTracking/KeyboardTrackingView';
+import KeyboardTrackingView, {KeyboardTrackingViewProps} from './KeyboardTracking/KeyboardTrackingView';
 import KeyboardAwareInsetsView from './KeyboardTracking/KeyboardAwareInsetsView';
 import KeyboardRegistry from './KeyboardInput/KeyboardRegistry';
-import KeyboardAccessoryView from './KeyboardInput/KeyboardAccessoryView';
+import KeyboardAccessoryView, {KeyboardAccessoryViewProps} from './KeyboardInput/KeyboardAccessoryView';
 import KeyboardUtils from './KeyboardInput/utils/KeyboardUtils';
 
-export default {KeyboardTrackingView, KeyboardAwareInsetsView, KeyboardRegistry, KeyboardAccessoryView, KeyboardUtils};
+export {KeyboardTrackingViewProps, KeyboardAccessoryViewProps};
+
+export default {
+  KeyboardTrackingView,
+  KeyboardAwareInsetsView,
+  KeyboardRegistry,
+  KeyboardAccessoryView,
+  KeyboardUtils
+};
