+++ tsconfig.build.json
@@ -3,6 +3,6 @@
   "compilerOptions": {
     "emitDeclarationOnly": true
   },
-  "include": ["src/**/*", "typings/**/*"],
+  "include": ["src/**/*", "typings/**/*", /*"lib/components/**//*"*/],
   "exclude": [ "node_modules", "src/index.ts"]
 }
